refactor: Stop using custom LottieView
